fix wrong type in XWorkArea
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@4805 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/xforms/XWorkArea.C
@@ -332,7 +332,7 @@ int XWorkArea::work_area_handler(FL_OBJECT * ob, int event,
 {
 	static int x_old = -1;
 	static int y_old = -1;
-	static long scrollbar_value_old = -1;
+	static double scrollbar_value_old = -1.0;
 
 	XEvent * ev = static_cast<XEvent*>(xev);
 	XWorkArea * area = static_cast<XWorkArea*>(ob->u_vdata);
